refactor: 영속성 추가

+++ src/main/java/me/yghee/codesharingplatform/domain/CodeNotFoundException.java
@@ -1,7 +1,7 @@
 package me.yghee.codesharingplatform.domain;
 
 public class CodeNotFoundException extends RuntimeException {
-    public CodeNotFoundException(int id) {
+    public CodeNotFoundException(Long id) {
         super("Could not find code "+id);
     }
 }
